BlackJack para cuatro jugadores
Esta versión es casi definitiva, solo falta añadir mas cartas a los jugadores.
Las mejoras que se han realizado han sido: 
- Añadir mas barajas a la partida para elevar la velocidad del juego.
- Cambiar la semilla de los números aleatorios ya que siempre se repetían.
- Mover las funciones a otro archivo cpp para dejar mejor ordenado el programa
- Añadir un sistema de turnos entre los jugadores ademas de una variable de bancarrota por si uno de los jugadores se quedara sin saldo no permitirle jugar. También si todos los jugadores se quedan sin saldo aparecerá un texto de END GAME que impedirá jugar a no ser que se salga o se reinicio la partida.
