+++ BlackJack-4Jugadores/BlackJack/BlackJack.pro
@@ -16,6 +16,7 @@ DEFINES += QT_DEPRECATED_WARNINGS
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
+    funciones.cpp \
     main.cpp \
     mainwindow.cpp \
     mano.cpp \
