+++ BlackJack-4Jugadores/BlackJack/BlackJack.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 4.11.0, 2020-04-28T18:34:11. -->
+<!-- Written by QtCreator 4.11.0, 2020-04-30T18:15:38. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
